Move the rendering out of the ECS
Remove the RenderAll system and move its functionality into main. This
was done in order to simplify creating "welcome" and "game over"
screens, which was complicated by the renderer being owned by the
RenderAll system.

The functionality shoud probably be moved out of main again, but not
into a system.